back/requirements.txt
amqp==2.1.1
billiard==3.5.0.2
celery==4.0.0
Django==1.10.3
django-cors-headers==1.3.1
django-debug-panel==0.8.3
django-debug-toolbar==1.6
django-filter==0.15.3
django-jenkins==0.110.0
djangorestframework==3.5.3
factory-boy==2.7.0
fake-factory==0.7.2
kombu==4.0.0
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
sqlparse==0.2.2
vine==1.1.3
